Moved more functionality over.
