+++ lib/authorizehelpers.js
@@ -45,6 +45,29 @@ function installDust(dust, db, query) {
       }
     }
   };
+
+  dust.helpers.requireUser = function(chunk, context, bodies, params) {
+    var user = context.get('user');
+    if (user) {
+      return chunk.render(bodies.block, context);
+    } else {
+      if (bodies.else) {
+        return chunk.render(bodies.else, context);
+      }
+    }
+  };
+
+  dust.helpers.isThisUser = function(chunk, context, bodies, params) {
+    var user = context.get('userPath');
+    var path = context.get('path');
+    if (user) {
+      if (user.toDottedPath() == path.toDottedPath()) {
+        return chunk.render(bodies.block, context);
+      } else {
+        return chunk.render(bodies.else, context);
+      }
+    }
+  };
 }
 
 exports.installDust = installDust;
